Refactor(db): pool 관리 코드를 별도의 py 파일로 분리
