+++ repository/news_channel_repository.py
@@ -1,0 +1,8 @@
+async def save_channel_games(pool, channel_id: int, games: list[str]):
+    pass
+
+async def delete_channel_games(pool, channel_id: int):
+    pass
+
+async def load_all_channel_games(pool):
+    pass
